+++ api/realtime.ts
@@ -1,0 +1,1 @@
+export { default } from '../services/ms-c-realtime/src/server.ts';